+++ lilya/context.py
@@ -10,6 +10,7 @@ from lilya.exceptions import ImproperlyConfigured
 from lilya.types import Doc, Scope
 
 if TYPE_CHECKING:
+    from lilya._internal._connection import Connection
     from lilya.requests import Request
     from lilya.routing import BasePath
 
@@ -390,19 +391,21 @@ class SessionContext:
     A context manager for handling session-specific request data using context variables.
     """
 
-    _session_context: contextvars.ContextVar[Request] = contextvars.ContextVar("session_context")
+    _session_context: contextvars.ContextVar[Connection] = contextvars.ContextVar(
+        "session_context"
+    )
 
     # Lazily access the session object
     lazy_session = LazySessionProxy(lambda: SessionContext._get_session_data())
 
     @classmethod
-    def set_request(cls, request: Request) -> None:
-        """Set the session context for the request."""
-        cls._session_context.set(request)
+    def set_connection(cls, connection: Connection) -> contextvars.Token:
+        """Set the session context for the Connection."""
+        return cls._session_context.set(connection)
 
     @classmethod
-    def get_request(cls) -> Request:
-        """Retrieve the current request (session context)."""
+    def get_connection(cls) -> Connection:
+        """Retrieve the current Connection (session context)."""
         try:
             return cls._session_context.get()
         except LookupError:
@@ -413,8 +416,8 @@ class SessionContext:
     @classmethod
     def _get_session_data(cls) -> Any:
         """Get the session data from the request."""
-        request = cls.get_request()
-        return getattr(request, "session", {})
+        connection = cls.get_connection()
+        return getattr(connection, "session", {})
 
     @classmethod
     def reset_context(cls, token: Any) -> None: